A dual-population algorithm based on alternative evolution and degeneration for solving constrained multi-objective optimization problems
It is challenging to solve constrained multi-objective optimization problems (CMOPs). Different from the traditional multi-objective optimization problem, the feasibility, convergence, and diversity of the population must be considered in the optimization process of a CMOP. How these factors are balanced will affect the performance of the constrained multi-objective optimization algorithm. To solve this problem, we propose a dual-population multi-objective optimization evolutionary algorithm. The proposed algorithm can make good use of its secondary population and alternative between evolution and degeneration according to the state of the secondary population to provide better information for the main population. The test results of three benchmark constrained multi-objective optimization problem suites, and four real-world constrained multi-objective optimization problems show that the algorithm is better than existing dual-population multi-objective optimization, especially when there is a distance between the unconstrained PF and the constrained PF.